+++ Common/Camera/ScreenShake.cs
@@ -1,21 +1,20 @@
-﻿using Microsoft.Xna.Framework;
-using TerrariaOverhaul.Utilities;
-
-namespace TerrariaOverhaul.Common.Camera;
+﻿namespace TerrariaOverhaul.Common.Camera;
 
 public struct ScreenShake
 {
+	public delegate float PowerDelegate(float progress);
+
 	public const float DefaultRange = 512f;
 
 	public float Power = 0.5f;
-	public Gradient<float>? PowerGradient;
+	public PowerDelegate? PowerFunction;
 	public float LengthInSeconds = 0.5f;
 	public float Range = DefaultRange;
 	public string? UniqueId;
 
-	public ScreenShake(Gradient<float> powerGradient, float lengthInSeconds) : this()
+	public ScreenShake(PowerDelegate powerFunction, float lengthInSeconds) : this()
 	{
-		PowerGradient = powerGradient;
+		PowerFunction = powerFunction;
 		LengthInSeconds = lengthInSeconds;
 	}