+++ Common/Magic/MagicWeapon.cs
@@ -1,4 +1,6 @@
-﻿using Terraria;
+﻿using System;
+using Microsoft.Xna.Framework;
+using Terraria;
 using Terraria.Audio;
 using Terraria.ID;
 using Terraria.ModLoader;
@@ -76,8 +78,17 @@ public partial class MagicWeapon : ItemOverhaul
 		});
 
 		if (!Main.dedServ) {
+			static float ScreenShakePowerFunction(float progress)
+			{
+				const float StartOffset = 0.05f;
+				const float MaxPower = 0.3f;
+				const float PowX = 3.0f;
+
+				return MathHelper.Clamp((MathF.Pow(progress, PowX) * (1f + StartOffset)) - StartOffset, 0f, 1f) * MaxPower;
+			}
+
 			item.EnableComponent<ItemPowerAttackScreenShake>(c => {
-				c.ScreenShake = new ScreenShake(chargeScreenShakePowerGradient, 0.5f);
+				c.ScreenShake = new ScreenShake(ScreenShakePowerFunction, float.PositiveInfinity);
 			});
 
 			item.EnableComponent<ItemPowerAttackSounds>(c => {