+++ CHANGELOG.md
@@ -49,6 +49,7 @@
 - Blood/Gel/Whatever particles' line lengths will now be rounded in rendering, with positions aligned to the pixel grid, appearing slightly more pixelated and low framerate, and preventing very thin lines forming at low velocity. They are still rotated however.
 - Crosshairs will now take "re-use delay" into account for their rotations.
 - Wooden arrow pieces will now be excepted from the "Gore Stay" feature, always disappearing in a short timespan.
+- Improved Magic Weapons' power attack charge screenshake.
 - Improved workarounds for FAudio's audio filtering issues:
 	- Using uncommon audio configurations will now only disable Reverb and not Low Pass Filtering.
 	- Reverb will now be disabled with unsupported audio device channel counts.